requirements-win-py312.txt
ray[rllib]==2.51.0
gymnasium==1.1.1
python-dotenv==1.0.1
networkx>=3.2
# Optional: enable TensorFlow-based path scoring if available
tensorflow==2.16.1
keras==3.4.1
